+++ earthdaily/earthdatastore/cube_utils/_zonal.py
@@ -9,7 +9,8 @@ from rasterio import features
 import numpy as np
 import xarray as xr
 import tqdm
-from . import custom_operations
+
+from . import custom_reducers
 from .preprocessing import rasterize
 from scipy.sparse import csr_matrix