+++ earthdaily/earthdatastore/cube_utils/geometry_manager.py
@@ -9,6 +9,9 @@ class GeometryManager:
         self.geometry = geometry
         self._obj = self.to_geopandas()
 
+    def __call__(self):
+        return self._obj
+
     def to_intersects(self, crs="EPSG:4326"):
         return json.loads(self._obj.to_crs(crs).dissolve().to_json(drop_id=True))[
             "features"
@@ -55,8 +58,12 @@ class GeometryManager:
                             [geometry["coordinates"][0]]
                         )
                         return gpd.GeoDataFrame(geometry=[geom], crs="EPSG:4326")
-                    elif isinstance(geometry, gpd.GeoSeries):
-                        self.input_type = "GeoSeries"
-                        return gpd.GeoDataFrame(geometry=geometry, crs="EPSG:4326")
+        elif isinstance(geometry, gpd.GeoSeries):
+            self.input_type = "GeoSeries"
+
+            return gpd.GeoDataFrame(
+                geometry=geometry,
+                crs="EPSG:4326" if geometry.crs is None else geometry.crs,
+            )
         else:
             raise NotImplementedError("Couldn't guess your geometry type")